Separated configuration logic
